bigints

+++ plugins/discourse-ai/db/migrate/20251120090000_add_llm_id_to_ai_api_audit_logs.rb
@@ -2,7 +2,7 @@
 
 class AddLlmIdToAiApiAuditLogs < ActiveRecord::Migration[8.0]
   def change
-    add_column :ai_api_audit_logs, :llm_id, :integer
+    add_column :ai_api_audit_logs, :llm_id, :bigint
     add_index :ai_api_audit_logs, [:llm_id]
   end
 end
